move meta.go to utils
